requirements.txt
torch>=2.0
mani_skill>=3.0.0b0
diffusers>=0.25
einops
wandb
hydra-core>=1.3
omegaconf
h5py
numpy
tqdm
